+++ src/config.rs
@@ -4,7 +4,6 @@
 use std::{collections::BTreeMap, ffi::OsStr};
 
 use anyhow::{format_err, Context as _, Result};
-use camino::Utf8Path;
 use serde::Deserialize;
 
 use crate::{env, term::Coloring};
@@ -25,18 +24,13 @@ pub(crate) struct Config {
 }
 
 impl Config {
-    pub(crate) fn new(
-        cargo: &OsStr,
-        workspace_root: &Utf8Path,
-        target: Option<&str>,
-        host: Option<&str>,
-    ) -> Result<Self> {
+    pub(crate) fn new(cargo: &OsStr, target: Option<&str>, host: Option<&str>) -> Result<Self> {
         // Use unstable cargo-config because there is no other good way.
         // However, it is unstable and can break, so allow errors.
         // https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#cargo-config
         // https://github.com/rust-lang/cargo/issues/9301
         let mut cargo = cmd!(cargo, "-Z", "unstable-options", "config", "get", "--format", "json");
-        cargo.dir(workspace_root).env("RUSTC_BOOTSTRAP", "1");
+        cargo.env("RUSTC_BOOTSTRAP", "1");
         let mut config = match cargo.read() {
             Ok(s) => serde_json::from_str(&s)
                 .with_context(|| format!("failed to parse output from {}", cargo))?,